+++ webapp/src/Settings.tsx
@@ -9,20 +9,27 @@ import {
   setTheme,
   setViewSize,
 } from "./settings";
+import { getClientPreferences, saveClientPreferences } from "./preferences";
 
 export function Settings() {
   return (
     <>
       <Top />
       <Container class={"mt-2"}>
-        <Row>
-          <Col></Col>
-          <Col md={8} sm={12} lg={6}>
+        <div class="row">
+          <div class="col"></div>
+          <div class="col-sm-12 col-md-8 col-lg-6">
             <Alert variant={"info"}>
               Note: Settings are stored client side and will not be reflected on
               other computers or in other browsers.
             </Alert>
+          </div>
+          <div class="col"></div>
+        </div>
 
+        <div class="row">
+          <div class="col"></div>
+          <div class="col-sm-12 col-md-8 col-lg-6">
             <div class={"row form-group"}>
               <label class="col-md-4 col-form-label">Theme</label>
               <div class="col-md-8">
@@ -42,9 +49,13 @@ export function Settings() {
                 </select>
               </div>
             </div>
+          </div>
+          <div class="col"></div>
+        </div>
 
-            <br />
-
+        <div class="row mt-2">
+          <div class="col"></div>
+          <div class="col-sm-12 col-md-8 col-lg-6">
             <div class={"row form-group"}>
               <label class="col-md-4 col-form-label">View Size</label>
               <div class="col-md-8">
@@ -73,9 +84,51 @@ export function Settings() {
                 </select>
               </div>
             </div>
-          </Col>
-          <Col></Col>
-        </Row>
+          </div>
+          <div class="col"></div>
+        </div>
+
+        <div class="row mt-2">
+          <div class="col"></div>
+          <div class="col-sm-12 col-md-8 col-lg-6">
+            <div class={"row form-group"}>
+              <label class="col-md-4 col-form-label">Timestamp Format</label>
+              <div class="col-md-8">
+                <select
+                  class="form-select"
+                  onchange={(e) => {
+                    let prefs = getClientPreferences();
+                    switch (e.currentTarget.value) {
+                      case "utc":
+                        prefs.timestamp_format = "utc";
+                        break;
+                      default:
+                        prefs.timestamp_format = undefined;
+                        break;
+                    }
+                    saveClientPreferences(prefs);
+                  }}
+                >
+                  <option
+                    value="local"
+                    selected={
+                      getClientPreferences().timestamp_format === "local"
+                    }
+                  >
+                    Local
+                  </option>
+                  <option
+                    value="utc"
+                    selected={getClientPreferences().timestamp_format === "utc"}
+                  >
+                    UTC
+                  </option>
+                </select>
+              </div>
+            </div>
+          </div>
+          <div class="col"></div>
+        </div>
       </Container>
     </>
   );